TST: Merge tests for solution strategy
